+++ test-data/src/main/java/org/ehrbase/test_data/operationaltemplate/OperationalTemplateTestData.java
@@ -109,7 +109,9 @@ public enum OperationalTemplateTestData {
       "conformance-ehrbase.de.v0"),
   MULTIMEDIA_TEST("MultimediaTest", "multimedia_test.en.v1.opt", "multimedia_test.en.v1"),
   NESTED("nested.en.v1", "nested.en.v1.opt", "nested.en.v1"),
+  RE_SPECT("ReSPECT-3.v0", "ReSPECT-3.v0.opt", "ReSPECT-3.v0"),
   IPS("International Patient Summary", "ips.v0.opt", "International Patient Summary"),
+  ERROR_TEST("Contains choice between interval and not interval", "ErrorTest.opt", "ErrorTest"),
   SECTION_CARDINALITY("Cardinality test template sections", "section_cardinality.opt", "cardinality_of_section");
 
   private final String filename;